feat(cells): cancel file upload [WPB-16843] (#18958)
* feat(cells): cancel file upload draft

* chore(deps): update @wireapp/core to version 46.20.12

* refactor(useFilePreview): extract cancelUpload logic into a separate function for improved readability

* refactor(fileUpload): streamline file upload cancellation logic and improve error handling

* chore(deps): update @wireapp/core to version 46.20.14 and @wireapp/api-client to version 27.37.1

* fix(FileCardLoading): correct progress calculation for loading styles

* refactor(FilePreviews): remove isLoading prop from preview components and streamline upload progress handling

* refactor(FilePreviewCard): update wrapperStyles to use CSSObject and simplify styling

* feat(InputBar): add FilePreviews component to display uploaded files in the input bar

* refactor(FilePreviewCard): remove unused styles from wrapperStyles

* chore: update yarn.lock

+++ src/script/components/InputBar/FilePreviews/useFilePreview/useFilePreview.ts
@@ -42,10 +42,21 @@ export const useFilePreview = ({file, cellsRepository, conversationQualifiedId}:
 
   const transformedName = isError ? `Upload failed: ${name}` : name;
 
+  const cancelUpload = () => {
+    cellsRepository.cancelUpload(file.id);
+    deleteFile({conversationId: conversationQualifiedId.id, fileId: file.id});
+  };
+
   const handleDelete = () => {
     if (file.preview) {
       URL.revokeObjectURL(file.preview);
     }
+
+    if (isLoading) {
+      cancelUpload();
+      return;
+    }
+
     deleteFile({conversationId: conversationQualifiedId.id, fileId: file.id});
     void cellsRepository.deleteFileDraft({uuid: file.remoteUuid, versionId: file.remoteVersionId});
   };
@@ -79,7 +90,6 @@ export const useFilePreview = ({file, cellsRepository, conversationQualifiedId}:
     name: transformedName,
     extension,
     size,
-    isLoading,
     isError,
     handleDelete,
     handleRetry,
